Fully encapsulate logic in `mro.mrs`
